Extract ScreenIndicator to util file. Remove duplication
