+++ src/components/ui/index.ts
@@ -2,3 +2,4 @@ export * from "./Button";
 export * from "./Input";
 export * from "./Text";
 export * from "./Alert";
+export * from "./ScreenIndicator";
