+++ src/screens/public/CreateAccount.tsx
@@ -15,15 +15,13 @@ import {
 import Animated, {
 	useAnimatedRef,
 	useAnimatedScrollHandler,
-	useAnimatedStyle,
 	useDerivedValue,
 	useSharedValue,
-	withTiming,
 } from "react-native-reanimated";
 import { SafeAreaView } from "react-native-safe-area-context";
 import * as z from "zod";
 
-import { Button, Input, Text, Alert } from "@/components/ui";
+import { Button, Input, Text, Alert, ScreenIndicator } from "@/components/ui";
 import { useAuth } from "@/hooks/useAuth";
 import tw from "@/lib/tailwind";
 import { PublicStackParamList } from "@/routes/public";
@@ -33,29 +31,6 @@ type CreateAccountProps = NativeStackScreenProps<
 >;
 
 const { width: SCREEN_WIDTH } = Dimensions.get("window");
-
-const ScreenIndicator = ({
-	activeIndex,
-	index,
-}: {
-	activeIndex: Animated.SharedValue<number>;
-	index: number;
-}) => {
-	const rIndicatorStyle = useAnimatedStyle(() => {
-		return {
-			width: withTiming(activeIndex.value === index ? 32 : 16),
-			backgroundColor: withTiming(
-				activeIndex.value === index ? "#4DAFFF" : "#D9D9D9",
-			),
-		};
-	});
-
-	return (
-		<Animated.View
-			style={[tw`h-[0.1875rem] bg-red-500 rounded-full`, rIndicatorStyle]}
-		/>
-	);
-};
 
 export function CreateAccount({ navigation }: CreateAccountProps) {
 	const scrollRef = useAnimatedRef<ScrollView>();
